Move avax chain and asset definitions
