+++ packages/xchain-avax/__e2e__/avax-client.e2e.ts
@@ -1,9 +1,9 @@
 import { Balance, Network, TxType } from '@xchainjs/xchain-client'
 import { ApproveParams, EstimateApproveParams, IsApprovedParams } from '@xchainjs/xchain-evm'
-import { Asset, AssetAVAX, assetAmount, assetToBase, assetToString } from '@xchainjs/xchain-util'
+import { Asset, assetAmount, assetToBase, assetToString } from '@xchainjs/xchain-util'
 
 import AvaxClient from '../src/client'
-import { defaultAvaxParams } from '../src/const'
+import { AssetAVAX, defaultAvaxParams } from '../src/util'
 
 // import { ApproveParams, EstimateApproveParams, IsApprovedParams } from '../src/types'
