+++ packages/xchain-avax/src/const.ts
@@ -1,101 +1,6 @@
-import { Network } from '@xchainjs/xchain-client'
-import { EVMClientParams, EtherscanProvider, ExplorerProvider } from '@xchainjs/xchain-evm'
-import { AssetAVAX } from '@xchainjs/xchain-util'
-import { BigNumber, ethers } from 'ethers'
+export const AVAX_DECIMAL = 8
 
-// =====Ethers providers=====
-const AVALANCHE_MAINNET_ETHERS_PROVIDER = new ethers.providers.JsonRpcProvider('https://api.avax.network/ext/bc/C/rpc')
-const AVALANCHE_TESTNET_ETHERS_PROVIDER = new ethers.providers.JsonRpcProvider(
-  'https://api.avax-test.network/ext/bc/C/rpc',
-)
+export const LOWER_FEE_BOUND = 2_000_000_000
+export const UPPER_FEE_BOUND = 1_000_000_000_000
 
-const ethersJSProviders = {
-  [Network.Mainnet]: AVALANCHE_MAINNET_ETHERS_PROVIDER,
-  [Network.Testnet]: AVALANCHE_TESTNET_ETHERS_PROVIDER,
-  [Network.Stagenet]: AVALANCHE_MAINNET_ETHERS_PROVIDER,
-}
-// =====Ethers providers=====
-// =====ONLINE providers=====
-
-const AVAX_ONLINE_PROVIDER_TESTNET = new EtherscanProvider(
-  AVALANCHE_TESTNET_ETHERS_PROVIDER,
-  'https://api-testnet.snowtrace.io',
-  'fake',
-  AssetAVAX.chain,
-  AssetAVAX,
-  18,
-)
-const AVAX_ONLINE_PROVIDER_MAINNET = new EtherscanProvider(
-  AVALANCHE_MAINNET_ETHERS_PROVIDER,
-  'https://api.snowtrace.io',
-  'fake',
-  AssetAVAX.chain,
-  AssetAVAX,
-  18,
-)
-const avaxProviders = {
-  [Network.Mainnet]: AVAX_ONLINE_PROVIDER_MAINNET,
-  [Network.Testnet]: AVAX_ONLINE_PROVIDER_TESTNET,
-  [Network.Stagenet]: AVAX_ONLINE_PROVIDER_MAINNET,
-}
-// =====ONLINE providers=====
-
-// =====Explorers=====
-const AVAX_MAINNET_EXPLORER = new ExplorerProvider(
-  'https://snowtrace.io/',
-  'https://snowtrace.io/address/%%ADDRESS%%',
-  'https://snowtrace.io/tx/%%TX_ID%%',
-)
-const AVAX_TESTNET_EXPLORER = new ExplorerProvider(
-  'https://testnet.snowtrace.io/',
-  'https://testnet.snowtrace.io/address/%%ADDRESS%%',
-  'https://testnet.snowtrace.io/tx/%%TX_ID%%',
-)
-const avaxExplorerProviders = {
-  [Network.Mainnet]: AVAX_MAINNET_EXPLORER,
-  [Network.Testnet]: AVAX_TESTNET_EXPLORER,
-  [Network.Stagenet]: AVAX_MAINNET_EXPLORER,
-}
-// =====Explorers=====
-
-const ethRootDerivationPaths = {
-  [Network.Mainnet]: `m/44'/60'/0'/0/`,
-  [Network.Testnet]: `m/44'/60'/0'/0/`,
-  [Network.Stagenet]: `m/44'/60'/0'/0/`,
-}
-
-const defaults = {
-  [Network.Mainnet]: {
-    approveGasLimit: BigNumber.from(200000),
-    transferGasAssetGasLimit: BigNumber.from(23000),
-    transferTokenGasLimit: BigNumber.from(100000),
-    gasPrice: BigNumber.from(30),
-  },
-  [Network.Testnet]: {
-    approveGasLimit: BigNumber.from(200000),
-    transferGasAssetGasLimit: BigNumber.from(23000),
-    transferTokenGasLimit: BigNumber.from(100000),
-    gasPrice: BigNumber.from(30),
-  },
-  [Network.Stagenet]: {
-    approveGasLimit: BigNumber.from(200000),
-    transferGasAssetGasLimit: BigNumber.from(23000),
-    transferTokenGasLimit: BigNumber.from(100000),
-    gasPrice: BigNumber.from(30),
-  },
-}
-export const defaultAvaxParams: EVMClientParams = {
-  chain: AssetAVAX.chain,
-  gasAsset: AssetAVAX,
-  gasAssetDecimals: 18,
-  defaults,
-  providers: ethersJSProviders,
-  explorerProviders: avaxExplorerProviders,
-  dataProviders: avaxProviders,
-  network: Network.Mainnet,
-  feeBounds: {
-    lower: 2_000_000_000,
-    upper: 1_000_000_000_000,
-  },
-  rootDerivationPaths: ethRootDerivationPaths,
-}
+export const AVAX_GAS_ASSET_DECIMAL = 18
