+++ packages/xchain-avax/src/index.ts
@@ -1,2 +1,3 @@
 export * from './client'
 export * from './const'
+export * from './util'
