+++ packages/xchain-thorchain-query/__e2e__/thorchain-liquidityP-e2e.ts
@@ -1,6 +1,6 @@
+import { AssetAVAX } from '@xchainjs/xchain-avax'
 import { Network } from '@xchainjs/xchain-client'
 import {
-  AssetAVAX,
   AssetBCH,
   AssetBNB,
   AssetBTC,
